Add Title and display in TOC
Reformat the list at the top of each page to be an asterisk list and to
make the labels bold. Add the RFC and Title headers and tweak the names
of the others.

Add an h2 line for the title to the template and make all other headers
h2. Remove the redundant anchor IDs.

Teach `generate-book.py` to read the title from the files and use it in
the TOC. Fall back on the file name if no title can be found.